Fix browse recipe filter feature
The recipe adapter was backed by the allRecipes list, which meant
clearing it would cause our cache of all recipes to be cleared.

Instead made the adapter have a copy of our recipes, so we are able
to clear it and repopulate it from the all recipe list with the respective
entries.